+++ src/lib/botClients/LexClient.ts
@@ -1,4 +1,4 @@
-import LexRuntime from 'aws-sdk/clients/lexruntime';
+import { LexRuntimeServiceClient, PostTextCommand } from '@aws-sdk/client-lex-runtime-service';
 import get from 'lodash.get';
 import { BotClient } from './BotClient';
 
@@ -8,8 +8,7 @@ export default class LexClient extends BotClient {
     private userId: string;
     private lastResponse: any;
     private sessionAttributes: any;
-    private props: any;
-    private lex: LexRuntime;
+    private lex: LexRuntimeServiceClient;
 
     constructor(botContext: any, userContext: any) {
         super(botContext, userContext);
@@ -19,29 +18,33 @@ export default class LexClient extends BotClient {
         this.lastResponse = null;
         this.sessionAttributes = this.userContext.userAttributes;
 
-        this.props = {
+        const clientConfig: any = {
             region: this.botContext.region,
         };
         // Optional Auth Environment Variables
-        this.props.accessKeyId = process.env.chatpickle_access_id || undefined;
-        this.props.secretAccessKey = process.env.chatpickle_access_secret || undefined;
+        if (process.env.chatpickle_access_id) {
+            clientConfig.credentials = {
+                accessKeyId: process.env.chatpickle_access_id,
+                secretAccessKey: process.env.chatpickle_access_secret,
+            };
+        }
 
-        this.lex = new LexRuntime(this.props);
+        this.lex = new LexRuntimeServiceClient(clientConfig);
         console.log(`[${this.userId}] New Conversation with ${this.botName}`);
     }
 
     public async speak(inputText: string): Promise<string> {
         console.log(`[${this.userId}] User: ${inputText}`);
 
-        const params = {
+        const command = new PostTextCommand({
             botName: this.botName,
             botAlias: this.botAlias,
             userId: this.userId,
             inputText,
             sessionAttributes: this.sessionAttributes,
-        };
+        });
 
-        this.lastResponse = await this.lex.postText(params).promise();
+        this.lastResponse = await this.lex.send(command);
         this.sessionAttributes = this.lastResponse.sessionAttributes;
 
         const reply: string = this.lastResponse.message.trim();